restore generated sdk/openapi/docs artifacts to merge-base

+++ others/python-sdk/src/autumn_sdk/models/batchtrackop.py
@@ -90,7 +90,7 @@ class RequestBodyTypedDict(TypedDict):
     r"""The amount of usage to record. Defaults to 1. Use negative values to credit balance (e.g., when removing a seat)."""
     properties: NotRequired[Dict[str, Any]]
     r"""Additional properties to attach to this usage event."""
-    timestamp: NotRequired[int]
+    timestamp: NotRequired[float]
     r"""Unix timestamp in milliseconds to use for the usage event. Defaults to the current time."""
     async_: NotRequired[bool]
     r"""If true, enqueue the event for asynchronous processing and return 204 immediately. The response will not include balance information."""
@@ -116,7 +116,7 @@ class RequestBody(BaseModel):
     properties: Optional[Dict[str, Any]] = None
     r"""Additional properties to attach to this usage event."""
 
-    timestamp: Optional[int] = None
+    timestamp: Optional[float] = None
     r"""Unix timestamp in milliseconds to use for the usage event. Defaults to the current time."""
 
     async_: Annotated[Optional[bool], pydantic.Field(alias="async")] = None
